api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "fastapi>=0.116.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
    "passlib[bcrypt]",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "bcrypt==4.1.2",
]
